Removed lwba router. Changed Adapter to PriceAdapter
