+++ packages/sources/coinmetrics-test/src/endpoint/index.ts
@@ -1,2 +1,2 @@
 export { endpoint as priceRouter } from './price-router'
-export { endpoint as lwbaRouter } from './lwba-router'
+export { endpoint as lwbaRouter } from './lwba-ws'
