+++ packages/sources/coinmetrics-test/src/index.ts
@@ -1,9 +1,9 @@
 import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
-import { Adapter } from '@chainlink/external-adapter-framework/adapter'
+import { PriceAdapter } from '@chainlink/external-adapter-framework/adapter'
 import { priceRouter, lwbaRouter } from './endpoint'
 import { config } from './config'
 
-export const adapter = new Adapter({
+export const adapter = new PriceAdapter({
   defaultEndpoint: 'price-ws',
   name: 'COINMETRICS',
   endpoints: [priceRouter, lwbaRouter],
